Move graphql queries, mutations and subscriptions closer to components
