+++ web/app/js/components/apollo/Mutation/Bookmarks/Unbookmark.tsx
@@ -1,6 +1,6 @@
 import { Mutation } from "react-apollo";
 import { Document } from "../../../../types/document";
-import mutation from "../../../../services/apollo/mutation/bookmarks/unbookmark.graphql";
+import mutation from "../../graphql/mutation/bookmarks/unbookmark.graphql";
 
 interface Data {
   bookmarks: {
