+++ web/app/js/components/apollo/Mutation/Syndication/Source.tsx
@@ -1,6 +1,6 @@
 import { Mutation } from "react-apollo";
 import { Source } from "../../../../types/syndication";
-import mutation from "../../../../services/apollo/mutation/syndication/source.graphql";
+import mutation from "../../graphql/mutation/syndication/source.graphql";
 
 interface Data {
   syndication: {
