+++ web/app/js/components/apollo/Mutation/Syndication/Status.tsx
@@ -1,7 +1,7 @@
 import { Mutation } from "react-apollo";
 import { Source } from "../../../../types/syndication";
-import enableSourceMutation from "../../../../services/apollo/mutation/syndication/enable-source.graphql";
-import disableSourceMutation from "../../../../services/apollo/mutation/syndication/disable-source.graphql";
+import enableSourceMutation from "../../graphql/mutation/syndication/enable-source.graphql";
+import disableSourceMutation from "../../graphql/mutation/syndication/disable-source.graphql";
 
 interface Data {
   syndication: {
