+++ web/app/js/components/apollo/Query/Feed.tsx
@@ -6,9 +6,9 @@ import {
 } from "apollo-boost";
 import { FeedVariables, FeedQueryData } from "../../../types/feed";
 import { getDataKey } from "../helpers/feed";
-import queryNews from "../../../services/apollo/query/feeds/news.graphql";
-import queryReadingList from "../../../services/apollo/query/feeds/reading-list.graphql";
-import queryFavorites from "../../../services/apollo/query/feeds/favorites.graphql";
+import queryNews from "../graphql/query/feeds/news.graphql";
+import queryReadingList from "../graphql/query/feeds/reading-list.graphql";
+import queryFavorites from "../graphql/query/feeds/favorites.graphql";
 
 export type FetchMore = <K extends keyof FeedVariables>(
   fetchMoreOptions: FetchMoreQueryOptions<FeedVariables, K> &
@@ -40,31 +40,31 @@ function getFetchMore(
   return data.feeds[key].results.length === data.feeds[key].total
     ? undefined
     : () =>
-        fetchMore({
-          variables: {
-            pagination: { from: data ? data.feeds[key].last : "" }
-          },
-          updateQuery: (prev, { fetchMoreResult: next }) => {
-            if (!next) {
-              return prev;
-            }
-
-            return {
-              feeds: {
-                ...prev.feeds,
-                [key]: {
-                  ...prev.feeds[key],
-                  last: next.feeds[key].last,
-                  limit: next.feeds[key].limit,
-                  results: [
-                    ...prev.feeds[key].results,
-                    ...next.feeds[key].results
-                  ]
-                }
-              }
-            };
+      fetchMore({
+        variables: {
+          pagination: { from: data ? data.feeds[key].last : "" }
+        },
+        updateQuery: (prev, { fetchMoreResult: next }) => {
+          if (!next) {
+            return prev;
           }
-        });
+
+          return {
+            feeds: {
+              ...prev.feeds,
+              [key]: {
+                ...prev.feeds[key],
+                last: next.feeds[key].last,
+                limit: next.feeds[key].limit,
+                results: [
+                  ...prev.feeds[key].results,
+                  ...next.feeds[key].results
+                ]
+              }
+            }
+          };
+        }
+      });
 }
 
 export {
