+++ web/app/js/components/apollo/graphql/fragment/bookmark.graphql
