+++ web/app/js/components/apollo/graphql/fragment/image.graphql
