+++ web/app/js/components/apollo/graphql/mutation/bookmarks/bookmark.graphql
