+++ web/app/js/components/apollo/graphql/mutation/bookmarks/create.graphql
