+++ web/app/js/components/apollo/graphql/mutation/bookmarks/favorite.graphql
