+++ web/app/js/components/apollo/graphql/mutation/bookmarks/unfavorite.graphql
