+++ web/app/js/components/apollo/graphql/mutation/syndication/delete-source.graphql
