+++ web/app/js/components/apollo/graphql/query/feeds/favorites.graphql
