+++ web/app/js/components/apollo/graphql/query/feeds/latest-news.graphql
