+++ web/app/js/components/apollo/graphql/query/feeds/news.graphql
