+++ web/app/js/components/apollo/graphql/subscription/favorites.graphql
