+++ web/app/js/components/apollo/graphql/subscription/news.graphql
